corrected comments and unit test thread src and dst address to 32bit

+++ unit_test_dma.c
@@ -8,8 +8,8 @@
 
 /* helper to run firmware_start_dma in separate thread */
 struct thread_args {
-    uint64_t src;
-    uint64_t dst;
+    uint32_t src;
+    uint32_t dst;
     uint32_t size;
     int result;
 };
